power: qpnp-charger: fix corner case for JEITA
When the ADC notifier is called there is a possibility
that the measured temperature is exactly equal to the
threshold rather than larger or smaller. In this case
JEITA rules are not applied.

Fix this by fixing the JEITA logic to also include
the exact threshold value and thus applying the
correct JEITA logic.

CRs-Fixed: 620291
Change-Id: Ib9462e7d28f6534c359873487420ec21b3e8aaa1
Signed-off-by: David Keitel <dkeitel@codeaurora.org>